* update setup for windows

+++ lib/model/nms/nms_kernel.cu
@@ -88,7 +88,7 @@ void _set_device(int device_id) {
   CUDA_CHECK(cudaSetDevice(device_id));
 }
 
-void _nms(int* keep_out, int* num_out, const float* boxes_host, int boxes_num,
+void _nms(long* keep_out, int* num_out, const float* boxes_host, int boxes_num,
           int boxes_dim, float nms_overlap_thresh, int device_id) {
   _set_device(device_id);
 
@@ -142,3 +142,5 @@ void _nms(int* keep_out, int* num_out, const float* boxes_host, int boxes_num,
   CUDA_CHECK(cudaFree(boxes_dev));
   CUDA_CHECK(cudaFree(mask_dev));
 }
+
+#include "gpu_nms.cpp"
